Revert "Revert "Revert "Decouple code and reorganize files."""
