+++ searchlib/src/vespa/searchlib/docstore/CMakeLists.txt
@@ -14,9 +14,7 @@ vespa_add_library(searchlib_docstore OBJECT
     idocumentstore.cpp
     logdatastore.cpp
     logdocumentstore.cpp
-    randreaders.cpp
     storebybucket.cpp
-    summaryexceptions.cpp
     visitcache.cpp
     writeablefilechunk.cpp
     DEPENDS
