+++ searchlib/src/vespa/searchlib/docstore/logdatastore.cpp
@@ -8,7 +8,6 @@
 #include <vespa/vespalib/data/fileheader.h>
 #include <vespa/vespalib/stllike/hash_map.hpp>
 #include <vespa/searchlib/common/rcuvector.hpp>
-#include <vespa/vespalib/util/exceptions.h>
 #include <thread>
 
 #include <vespa/log/log.h>
